drm/xe: Implement VM snapshot support for BO's and userptr, v2.
Since we cannot immediately capture the BO's and userptr, perform it in
2 stages. The immediate stage takes a reference to each BO and userptr,
while a delayed worker captures the contents and then frees the
reference.

This is required because in signaling context, no locks can be taken, no
memory can be allocated, and no waits on userspace can be performed.

With the delayed worker, all of this can be performed very easily,
without having to resort to hacks.

Changes since v1:
- Fix crash on NULL captured vm.
- Use ascii85_encode to capture BO contents and save some space. (Jose
  Souza)
- Add length to coredump output for each captured area.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>